kilimanjaro/pyproject.toml
[tool.poetry]
name = "kilimanjaro"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.6"
python-decouple = "3.4"
channels = "3.0.4"
channels-redis = "3.3.0"
psycopg2-binary = "2.9.1"
nplusone = "^1.0.0"
django-debug-toolbar = "^3.2.2"
django-seal = "^1.4.4"
django-model-utils = "^4.1.1"
hashids = "^1.3.1"
phonenumbers = "^8.12.31"
Pillow = "^8.3.2"
django-money = "^2.0.3"
djangorestframework = "^3.12.4"
requests = "^2.26.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.2.0"}
djangorestframework-simplejwt = "^5.0.0"
django-redis = "^5.0.0"
django-filter = "^2.4.0"
django-import-export = "^2.6.0"
openpyxl = "^3.0.9"
pandas = "^1.3.3"
celery = "^5.1.2"
redis = "^3.5.3"
mutagen = "^1.45.1"


[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
isort = "^5.9.3"
black = "^21.9b0"
pytest-django = "^4.4.0"
pytest-postgresql = "^3.1.2"
mixer = "^7.2.0"
coverage = "^6.1"
pytest-asyncio = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
